src/ChunkProcessor/requirements.txt
sentence-transformers
numpy
hf_xet
psycopg2-binary>=2.9.3